The logical-schema XML file was changed to contain the forestTemp extent which is required by the refactoring example.
The RefactorerTest has been introduced to test in JUnit how it behaves. There seems to be a problem in the *translation* phase. Everything else is working smoothly.

Changed the QueryCompiler class to take into account the added exceptions and the interfaces that have been requested by the data analysis module.