Fixed the translator bug where nested queries were allowed to not be given a name. At this point, only sub-queries that have a reference identifier are allowed. Unreferenced named queries are not allowed at this point.
Also, another bug has been fixed, where aliasing the same attribute twice resulted in it not being recognized.
This in turn produced an error, saying that an attribute was missing.

Added multiple Unit Tests to check that the functionality is now correct. These can be
found in the TranslatorTest.java file, at the bottom.

In the logical-schema.xml file of the compiler module, in the /src/test/resources/etc, the name of the columns
for integer values is integetColumn. This has been changed to integerColumn.

*Notice*: The main method of the test client uses other logical-schema.xml files to read the extent (and
sources) metadata. This results in inconsistencies and the test cases (junits) need to be re-written. Perhaps,
change it to a single logical schema file.