Current modifications for the successor relation. Currently got output code generated (not tested) Path from client to planner starting successor relation created and tested (used with new SNEE props values, hard coded currently to allow successor, can be changed in default props). Alternative QEPs are generated successfully. Required to fix a bug in the DAF generator where it places exchanges into the PAF. these are now removed right after the DAF is generated. TABU search code is just a skeleton. Moved the whenSchedulerException into the API from the compiler. as now its an exception the client may need to deal with. Tweaked tests in accordance with the move of whenSchedulerException.
Tweaked the meta routers heuristics so that they work for all when generating alternative plans, and only use 1 when working with the dynamic strategy. 
Added supporting code to allow the meta router to be used to generate full RT's without the dependence on a fragmented RT.

successor class contains a QEP and a Agenda count for the tabu search. May also add a copy of the running sites for the TABU search to operate over time also.