tweaks to allow overlay to search all nodes when one node fails to produce an adaptation, now applicable to all strategies.
fixed the code generator which removed wirings from consideration based off sites comparision and not the sites ids (linked to now that we have possible clones of sites which are only equivlant by ids. 

added tweaks to the transfer of qeps onto overlay sites so that they are recorded better in agendas, and that the agendas only output sites based from the rt currently running.